Added account total values
