+++ finance-dashboard-frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import LoginPage from "./pages/LoginPage";
 import RegisterPage from "./pages/RegisterPage";
-import DashboardPage from "./pages/DashboardPage";
+import DashboardPage from "./pages/Dashboard/DashboardPage";
 import TransactionsPage from "./pages/TransactionsPage";
 import ProtectedRoute from "./auth/ProtectedRoute";
